Refactor AI filename recommendation functionality and move to lib/ai.py
